Upgrading DL4J: Not working yet
I'm trying to upgrade the DL4J version. The code Maven builds, but code that requires DL4J doesn't work yet. I'm focusing on the PictureInnovationTask and the NeuralStyleTransfer.

It seems that the code CAN now successfully download the VGG models, so that solves #583 ... maybe.

I should be able to get NST to work #158 since a working version of my code is included in the latest version of DL4J! I am still getting the weird error though:
Cannot perform backprop in layer ConvolutionLayer

+++ src/main/java/edu/southwestern/util/graphics/GraphicsUtil.java
@@ -482,7 +482,7 @@ public class GraphicsUtil {
 	 * @return BufferedImage 
 	 */
 	public static BufferedImage imageFromINDArray(INDArray array) {
-		int[] shape = array.shape();
+		long[] shape = array.shape();
 		// Should the order of these be switched?
 		int width = (int) shape[2];
 		int height = (int) shape[3];
